Add function diffdotfiles
Compares a file with its corresponding version in the `dotfiles/`
directory. Finds the file by mirroring the directory structure inside
`dotfiles/`, but also handles special cases for files outside of `~`,
symlinks which resolve to a file in `dotfiles`, and the files `.gitconfig`
and `.gitignore` (which conflict with the local versions in the dotfiles
repo and are thus stored under the `dotfiles/ROOT/home/troy/` subdirectory).