Define `lscomposefiles` as a function to list all compose files
The function looks for `include` statements in `~/.XCompose`, and extracts
the filenames (after expanding any % substitutions).

TODO:
-----
The files are listed in order, and `~/.XCompose` is added at the end.
Currently, this is accurate, since the includes all occur at the start of
`.XCompose`.

To handle includes that occur elsewhere, and keep the correct order,
`lscompose` would no longer be able to simply concatenate the files listed by
`lscomposefiles`. Instead, it could scan for include lines as it prints the
compose key combinations, expanding included files in place.